Move frontend.rs to a directory
